Rename and extract ImageUsage in its own module
