+++ vulkano/src/image/sys.rs
@@ -25,6 +25,7 @@ use device::Device;
 use format::Format;
 use format::FormatTy;
 use image::ImageDimensions;
+use image::ImageUsage;
 use image::MipmapsCount;
 use image::ViewType;
 use memory::DeviceMemory;
@@ -78,7 +79,7 @@ impl UnsafeImage {
     /// - Panics if the number of samples is 0.
     ///
     #[inline]
-    pub unsafe fn new<'a, Mi, I>(device: &Arc<Device>, usage: &Usage, format: Format,
+    pub unsafe fn new<'a, Mi, I>(device: &Arc<Device>, usage: &ImageUsage, format: Format,
                                  dimensions: ImageDimensions, num_samples: u32, mipmaps: Mi,
                                  sharing: Sharing<I>, linear_tiling: bool,
                                  preinitialized_layout: bool)
@@ -95,7 +96,7 @@ impl UnsafeImage {
     }
 
     // Non-templated version to avoid inlining and improve compile times.
-    unsafe fn new_impl(device: &Arc<Device>, usage: &Usage, format: Format,
+    unsafe fn new_impl(device: &Arc<Device>, usage: &ImageUsage, format: Format,
                        dimensions: ImageDimensions, num_samples: u32, mipmaps: MipmapsCount,
                        (sh_mode, sh_indices): (vk::SharingMode, SmallVec<[u32; 8]>),
                        linear_tiling: bool, preinitialized_layout: bool)
@@ -153,7 +154,7 @@ impl UnsafeImage {
         // If `transient_attachment` is true, then only `color_attachment`,
         // `depth_stencil_attachment` and `input_attachment` can be true as well.
         if usage.transient_attachment {
-            let u = Usage {
+            let u = ImageUsage {
                 transient_attachment: false,
                 color_attachment: false,
                 depth_stencil_attachment: false,
@@ -161,7 +162,7 @@ impl UnsafeImage {
                 .. usage.clone()
             };
 
-            if u != Usage::none() {
+            if u != ImageUsage::none() {
                 return Err(ImageCreationError::UnsupportedUsage);
             }
         }
@@ -972,120 +973,6 @@ impl Drop for UnsafeImageView {
     }
 }
 
-/// Describes how an image is going to be used. This is **not** an optimization.
-///
-/// If you try to use an image in a way that you didn't declare, a panic will happen.
-///
-/// If `transient_attachment` is true, then only `color_attachment`, `depth_stencil_attachment`
-/// and `input_attachment` can be true as well. The rest must be false or an error will be returned
-/// when creating the image.
-#[derive(Debug, Copy, Clone, PartialEq, Eq, Hash)]
-pub struct Usage {
-    /// Can be used a source for transfers. Includes blits.
-    pub transfer_source: bool,
-
-    /// Can be used a destination for transfers. Includes blits.
-    pub transfer_dest: bool,
-
-    /// Can be sampled from a shader.
-    pub sampled: bool,
-
-    /// Can be used as an image storage in a shader.
-    pub storage: bool,
-
-    /// Can be attached as a color attachment to a framebuffer.
-    pub color_attachment: bool,
-
-    /// Can be attached as a depth, stencil or depth-stencil attachment to a framebuffer.
-    pub depth_stencil_attachment: bool,
-
-    /// Indicates that this image will only ever be used as a temporary framebuffer attachment.
-    /// As soon as you leave a render pass, the content of transient images becomes undefined.
-    ///
-    /// This is a hint to the Vulkan implementation that it may not need allocate any memory for
-    /// this image if the image can live entirely in some cache.
-    pub transient_attachment: bool,
-
-    /// Can be used as an input attachment. In other words, you can draw to it in a subpass then
-    /// read from it in a following pass.
-    pub input_attachment: bool,
-}
-
-impl Usage {
-    /// Builds a `Usage` with all values set to true. Note that using the returned value will
-    /// produce an error because of `transient_attachment` being true.
-    #[inline]
-    pub fn all() -> Usage {
-        Usage {
-            transfer_source: true,
-            transfer_dest: true,
-            sampled: true,
-            storage: true,
-            color_attachment: true,
-            depth_stencil_attachment: true,
-            transient_attachment: true,
-            input_attachment: true,
-        }
-    }
-
-    /// Builds a `Usage` with all values set to false. Useful as a default value.
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// use vulkano::image::Usage as ImageUsage;
-    ///
-    /// let _usage = ImageUsage {
-    ///     transfer_dest: true,
-    ///     sampled: true,
-    ///     .. ImageUsage::none()
-    /// };
-    /// ```
-    #[inline]
-    pub fn none() -> Usage {
-        Usage {
-            transfer_source: false,
-            transfer_dest: false,
-            sampled: false,
-            storage: false,
-            color_attachment: false,
-            depth_stencil_attachment: false,
-            transient_attachment: false,
-            input_attachment: false,
-        }
-    }
-
-    #[doc(hidden)]
-    #[inline]
-    pub fn to_usage_bits(&self) -> vk::ImageUsageFlagBits {
-        let mut result = 0;
-        if self.transfer_source { result |= vk::IMAGE_USAGE_TRANSFER_SRC_BIT; }
-        if self.transfer_dest { result |= vk::IMAGE_USAGE_TRANSFER_DST_BIT; }
-        if self.sampled { result |= vk::IMAGE_USAGE_SAMPLED_BIT; }
-        if self.storage { result |= vk::IMAGE_USAGE_STORAGE_BIT; }
-        if self.color_attachment { result |= vk::IMAGE_USAGE_COLOR_ATTACHMENT_BIT; }
-        if self.depth_stencil_attachment { result |= vk::IMAGE_USAGE_DEPTH_STENCIL_ATTACHMENT_BIT; }
-        if self.transient_attachment { result |= vk::IMAGE_USAGE_TRANSIENT_ATTACHMENT_BIT; }
-        if self.input_attachment { result |= vk::IMAGE_USAGE_INPUT_ATTACHMENT_BIT; }
-        result
-    }
-
-    #[inline]
-    #[doc(hidden)]
-    pub fn from_bits(val: u32) -> Usage {
-        Usage {
-            transfer_source: (val & vk::IMAGE_USAGE_TRANSFER_SRC_BIT) != 0,
-            transfer_dest: (val & vk::IMAGE_USAGE_TRANSFER_DST_BIT) != 0,
-            sampled: (val & vk::IMAGE_USAGE_SAMPLED_BIT) != 0,
-            storage: (val & vk::IMAGE_USAGE_STORAGE_BIT) != 0,
-            color_attachment: (val & vk::IMAGE_USAGE_COLOR_ATTACHMENT_BIT) != 0,
-            depth_stencil_attachment: (val & vk::IMAGE_USAGE_DEPTH_STENCIL_ATTACHMENT_BIT) != 0,
-            transient_attachment: (val & vk::IMAGE_USAGE_TRANSIENT_ATTACHMENT_BIT) != 0,
-            input_attachment: (val & vk::IMAGE_USAGE_INPUT_ATTACHMENT_BIT) != 0,
-        }
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use std::iter::Empty;
@@ -1093,7 +980,7 @@ mod tests {
 
     use super::ImageCreationError;
     use super::UnsafeImage;
-    use super::Usage;
+    use super::ImageUsage;
 
     use image::ImageDimensions;
     use format::Format;
@@ -1103,9 +990,9 @@ mod tests {
     fn create_sampled() {
         let (device, _) = gfx_dev_and_queue!();
 
-        let usage = Usage {
+        let usage = ImageUsage {
             sampled: true,
-            .. Usage::none()
+            .. ImageUsage::none()
         };
 
         let (_img, _) = unsafe {
@@ -1120,10 +1007,10 @@ mod tests {
     fn create_transient() {
         let (device, _) = gfx_dev_and_queue!();
 
-        let usage = Usage {
+        let usage = ImageUsage {
             transient_attachment: true,
             color_attachment: true,
-            .. Usage::none()
+            .. ImageUsage::none()
         };
 
         let (_img, _) = unsafe {
@@ -1138,9 +1025,9 @@ mod tests {
     fn zero_sample() {
         let (device, _) = gfx_dev_and_queue!();
 
-        let usage = Usage {
+        let usage = ImageUsage {
             sampled: true,
-            .. Usage::none()
+            .. ImageUsage::none()
         };
 
         let res = unsafe {
@@ -1160,9 +1047,9 @@ mod tests {
     fn non_po2_sample() {
         let (device, _) = gfx_dev_and_queue!();
 
-        let usage = Usage {
+        let usage = ImageUsage {
             sampled: true,
-            .. Usage::none()
+            .. ImageUsage::none()
         };
 
         let res = unsafe {
@@ -1182,9 +1069,9 @@ mod tests {
     fn zero_mipmap() {
         let (device, _) = gfx_dev_and_queue!();
 
-        let usage = Usage {
+        let usage = ImageUsage {
             sampled: true,
-            .. Usage::none()
+            .. ImageUsage::none()
         };
 
         let res = unsafe {
@@ -1205,9 +1092,9 @@ mod tests {
     fn mipmaps_too_high() {
         let (device, _) = gfx_dev_and_queue!();
 
-        let usage = Usage {
+        let usage = ImageUsage {
             sampled: true,
-            .. Usage::none()
+            .. ImageUsage::none()
         };
 
         let res = unsafe {
@@ -1230,9 +1117,9 @@ mod tests {
     fn shader_storage_image_multisample() {
         let (device, _) = gfx_dev_and_queue!();
 
-        let usage = Usage {
+        let usage = ImageUsage {
             storage: true,
-            .. Usage::none()
+            .. ImageUsage::none()
         };
 
         let res = unsafe {
@@ -1253,9 +1140,9 @@ mod tests {
     fn compressed_not_color_attachment() {
         let (device, _) = gfx_dev_and_queue!();
 
-        let usage = Usage {
+        let usage = ImageUsage {
             color_attachment: true,
-            .. Usage::none()
+            .. ImageUsage::none()
         };
 
         let res = unsafe {
@@ -1276,10 +1163,10 @@ mod tests {
     fn transient_forbidden_with_some_usages() {
         let (device, _) = gfx_dev_and_queue!();
 
-        let usage = Usage {
+        let usage = ImageUsage {
             transient_attachment: true,
             sampled: true,
-            .. Usage::none()
+            .. ImageUsage::none()
         };
 
         let res = unsafe {
@@ -1299,9 +1186,9 @@ mod tests {
     fn cubecompatible_dims_mismatch() {
         let (device, _) = gfx_dev_and_queue!();
 
-        let usage = Usage {
+        let usage = ImageUsage {
             sampled: true,
-            .. Usage::none()
+            .. ImageUsage::none()
         };
 
         let res = unsafe {
