+++ vulkano/src/swapchain/swapchain.rs
@@ -30,8 +30,8 @@ use format::FormatDesc;
 use image::ImageAccess;
 use image::ImageDimensions;
 use image::ImageLayout;
+use image::ImageUsage;
 use image::sys::UnsafeImage;
-use image::sys::Usage as ImageUsage;
 use image::swapchain::SwapchainImage;
 use swapchain::ColorSpace;
 use swapchain::CompositeAlpha;
